% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-core.R
\name{chiblihash64}
\alias{chiblihash64}
\title{Calculate the 64-bit chibli hash for the given data}
\usage{
chiblihash64(raw_vec, skip = 0, len = length(raw_vec))
}
\arguments{
\item{raw_vec}{raw vector}

\item{skip}{number of bytes to skip at start}

\item{len}{Number of bytes to hash}
}
\value{
String containing the chibli hash
}
\description{
Calculate the 64-bit chibli hash for the given data
}
\examples{
zz <- as.raw(1:100)
chiblihash64(zz)
chiblihash64(zz)
zz[1] <- as.raw(199)
chiblihash64(zz)
}
