% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-core.R
\name{fec_prepare_file}
\alias{fec_prepare_file}
\title{Create blocks for a file}
\usage{
fec_prepare_file(
  filename,
  blocks_filename = NULL,
  k = 10,
  n = 2,
  verbosity = 0L
)
}
\arguments{
\item{filename}{filename}

\item{blocks_filename}{Default: NULL, suffix with ".feck"}

\item{k}{number of chunks to split data into. Larger 'k' will mean the file will
be chunked into more, smaller blocks.  This will help isolate errors,
but can increase computation time.}

\item{n}{number of repair chunks to create.  Each repair chunk will be 
able to repair one bad chunk in the data.}

\item{verbosity}{Default: 0}
}
\value{
Invisibly return the raw vector of repair blocks
}
\description{
Create blocks for a file
}
