% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-core.R
\name{fec_prepare_raw}
\alias{fec_prepare_raw}
\title{Create recovery blocks}
\usage{
fec_prepare_raw(raw_vec, k = 10, n = 2, verbosity = 0L)
}
\arguments{
\item{raw_vec}{raw vector}

\item{k}{number of chunks to split data into. Larger 'k' will mean the file will
be chunked into more, smaller blocks.  This will help isolate errors,
but can increase computation time.}

\item{n}{number of repair chunks to create.  Each repair chunk will be 
able to repair one bad chunk in the data.}

\item{verbosity}{Default: 0}
}
\value{
Raw vector of meta-data and recovery blocks
}
\description{
Create recovery blocks
}
\examples{
set.seed(1)
dat <- as.raw(sample(0:255, 1e3, replace = TRUE))
rblocks <- fec_prepare_raw(dat, verbosity = 1)
}
