% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-core.R
\name{fec_repair_file}
\alias{fec_repair_file}
\title{Repair file}
\usage{
fec_repair_file(filename, blocks_filename = NULL, verbosity = 0L)
}
\arguments{
\item{filename}{filename. Repaired file will have suffix ".repaired"}

\item{blocks_filename}{Default: NULL, suffix with ".feck"}

\item{verbosity}{Default: 0}
}
\value{
Invisibly return a raw vector of the contents of the repaired file.
        If repair is not possible, an error will be raised.
}
\description{
Repair file
}
