% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-core.R
\name{fec_repair_raw}
\alias{fec_repair_raw}
\title{Recover damaged data in a raw vector}
\usage{
fec_repair_raw(raw_vec, blocks, verbosity = 0L)
}
\arguments{
\item{raw_vec}{raw vector}

\item{blocks}{raw vector of recovery blocks as created by \code{\link{fec_prepare_raw}()}}

\item{verbosity}{Default: 0}
}
\value{
Repaired data, or NULL if repair not possible
}
\description{
Recover damaged data in a raw vector
}
\examples{
set.seed(1)
dat0 <- dat <- as.raw(sample(0:255, 1e4, replace = TRUE))
rblocks <- fec_prepare_raw(dat, verbosity = 1)
# simulate damange to data
dat[1:100] <- as.raw(0)
identical(dat, dat0)
dat_repaired <- fec_repair_raw(dat, rblocks, verbosity = 1)
identical(dat_repaired, dat0)
}
