% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-core.R
\name{print.fec_blocks}
\alias{print.fec_blocks}
\title{Print repair blocks}
\usage{
\method{print}{fec_blocks}(x, ...)
}
\arguments{
\item{x}{Repair blocks created by \code{\link{fec_prepare_raw}()}}

\item{...}{ignored}
}
\value{
None
}
\description{
Print repair blocks
}
\examples{
rblocks <- fec_prepare_raw(as.raw(1:1000))
print(rblocks)
}
